requirements.txt
-r requirements-dev.txt

psycopg2-binary==2.9.3
django-storages==1.12.3
django-storages-azure==1.6.8